Moved UITextView to a new file.
